refactor: moving `formatting` and `response` into lang
These are both Golang specific helpers so it makes sense to group these
